fix: resolve all 157 ESLint warnings and errors
- Extract LatestResponseCard and SessionOverviewCard components from
  SummaryTab.vue to bring it under the 600-line limit (681 → 305 lines)
- Add type definitions and default values to props in AddSessionToLaneModal,
  AutoRescheduleModal, ScheduleSessionModal, SchedulingEditModal,
  LaneSettingsModal
- Reformat ArchiveConfirmModal.vue template: one attribute per line,
  correct ordering (class/aria-label/:disabled before @click events),
  proper newlines for element content, void element self-closing
- Fix ConversationTab.vue: move @update:selectedModel after all bindings
  and hyphenate the event name to @update:selected-model
- Fix SystemIndicators.vue: indent 3 misaligned aria-hidden attributes
- Add eslint-disable/enable block comments around intentional v-html
  usages in ButtonStatusModal, CanvasFileViewer, CommandButtonItem,
  MarkdownViewer (syntax highlighting and markdown rendering)
- Extend test file ESLint overrides to turn off vue/one-component-per-file
  and vue/require-prop-types (inline mock components in tests are intentional)
- Fix ArchiveConfirmModal.test.js: use .trim() on textContent comparisons
  for buttons reformatted to multi-line
