+++ packages/web/src/components/ConversationTab.vue
@@ -52,7 +52,6 @@
       :save-status="saveStatus"
       :input-has-content="inputHasContent"
       :thinking-enabled="sessionsStore.currentSession?.thinkingEnabled"
-      @update:selectedModel="selectedModel = $event"
       :working-directory="workingDirectory"
       :project-id="sessionsStore.currentSession?.projectId"
       :current-template-id="sessionsStore.currentSession?.nextTemplateId"
@@ -62,6 +61,7 @@
       :send-button-disabled-reason="sendButtonDisabledReason"
       :is-send-disabled="isSendDisabled"
       :auto-send-pending-prompt="sessionsStore.currentSession?.autoSendPendingPrompt ?? false"
+      @update:selected-model="selectedModel = $event"
       @submit="handleFormSubmit"
       @auto-send-toggle="handleAutoSendToggle"
       @input="handleInput"
